New Version of the Hospital project
In this version I deleted the Hospital class because it was useless. Also, modified class Patient so it didn't just have one object medicine, and change it so it could have an entire list of medicine. Created a new class calles Appointment as a reference to those moments when the doctor needs to meet again with the patient to check out how is he or she going with the treatment. And at last, I modified class Expedient, so it wasn't like a second main class but a class thar have the Patient, Doctor and Nurse classes so it could return the expedient of the patient.
Those are mainly the changes I commited in this new version.